+++ backend/packages/provider-s3/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @cat-factory/provider-s3
 
+## 0.2.293
+
+### Patch Changes
+
+- @cat-factory/kernel@0.234.1
+
+## 0.2.292
+
+### Patch Changes
+
+- Updated dependencies [937d4af]
+  - @cat-factory/kernel@0.234.0
+
+## 0.2.291
+
+### Patch Changes
+
+- Updated dependencies [2580fee]
+- Updated dependencies [eb4ca17]
+  - @cat-factory/kernel@0.233.0
+
 ## 0.2.290
 
 ### Patch Changes